--- PodcastApp/PodcastApp.sln.old
@@ -1,25 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28010.2041
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "WindowsFormsApp1", "WindowsFormsApp1\WindowsFormsApp1.csproj", "{3BE727E2-FFD7-4D3A-AA34-0B1B26096932}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{3BE727E2-FFD7-4D3A-AA34-0B1B26096932}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{3BE727E2-FFD7-4D3A-AA34-0B1B26096932}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{3BE727E2-FFD7-4D3A-AA34-0B1B26096932}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{3BE727E2-FFD7-4D3A-AA34-0B1B26096932}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {B50DE0CA-0219-44CA-B344-3D5A92BB1DF2}
-	EndGlobalSection
-EndGlobal
