+++ PodcastApp/PodcastApp/.vs/VSWorkspaceState.json
@@ -1,0 +1,11 @@
+{
+  "ExpandedNodes": [
+    "",
+    "\\Business logics",
+    "\\DataLayer",
+    "\\Entities",
+    "\\GUI"
+  ],
+  "SelectedNode": "\\Form1.cs",
+  "PreviewInSolutionExplorer": false
+}
