+++ PodcastApp/PodcastApp/PodcastApp.csproj
@@ -46,7 +46,9 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="DataRetriever.cs" />
+    <Compile Include="Category.cs" />
+    <Compile Include="RssRetriever.cs" />
+    <Compile Include="DirectoryCreator.cs" />
     <Compile Include="Form1.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -56,6 +58,8 @@
     <Compile Include="Podcast.cs" />
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="StringManipulator.cs" />
+    <Compile Include="XmlReader.cs" />
     <Compile Include="XmlWriter.cs" />
     <EmbeddedResource Include="Form1.resx">
       <DependentUpon>Form1.cs</DependentUpon>
