+++ PodcastApp/PodcastApp/Program.cs
@@ -17,6 +17,9 @@ namespace WindowsFormsApp1
             Application.EnableVisualStyles();
             Application.SetCompatibleTextRenderingDefault(false);
             Application.Run(new Form1());
+
         }
+
+       
     }
 }
